+++ Changelog.md
@@ -27,9 +27,10 @@
 - Fix: The Sentry-/Mine Remote Access Tool screens do not show the item's custom name
 - Fix: Reinforced Doors can be closed by giving them a block update
 - Fix: Reinforced Doors that are next to, but don't face, each other can act as double doors
+- Fix: Certain blocks don't update their indirect neighbors properly when getting destroyed while in a powered state
 - Fix: The Reinforced Mangrove Button and Pressure Plate can't open reinforced doors
 
-* only available if the update_1_20 datapack is enabled
+[*] only available if the update_1_20 datapack is enabled
 
 --------------------------Changelog for v1.9.6.1 of SecurityCraft--------------------------